+++ Datax/UI.cpp
@@ -1,4 +1,6 @@
+
 #include"GameBaseData.h"
+
 #include "UI.h"
 
 
@@ -7,23 +9,28 @@ void UI::PrintScene_UI(BasicFunctions* bFunctions, uint UIscene)
 
     switch (UIscene)
     {
-    case 0:
+
+    case (uint)0:
 
         uiCase0(bFunctions);
+
         break;
 
-    case 1:
+    case (uint)1:
 
         uiCase1(bFunctions);
+
         break;
 
-    case 2:
+    case (uint)2:
 
         uiCase2(bFunctions);
+
         break;
 
 
     default:
+
         break;
 
     }
@@ -33,13 +40,19 @@ void UI::PrintScene_UI(BasicFunctions* bFunctions, uint UIscene)
 
 void UI::uiCase0(BasicFunctions* bFunctions)
 {
+
     bFunctions->SetCursor_UI_Left();
+
     bFunctions->PrintString("UI ONE");
 
+
     bFunctions->SetCursor_UI_Center();
+
     bFunctions->PrintString("UI TWO");
+
               
     bFunctions->SetCursor_UI_Right();
+
     bFunctions->PrintString("UI THREE");
 
 }
@@ -48,12 +61,17 @@ void UI::uiCase1(BasicFunctions* bFunctions)
 {
 
     bFunctions->SetCursor_UI_Left();
+
     bFunctions->PrintString("UI ONE");
+
               
     bFunctions->SetCursor_UI_Center();
+
     bFunctions->PrintString("UI TWO");
+
               
     bFunctions->SetCursor_UI_Right();
+
     bFunctions->PrintString("UI THREE");
 
 }
@@ -62,12 +80,17 @@ void UI::uiCase2(BasicFunctions* bFunctions)
 {
 
     bFunctions->SetCursor_UI_Left();
+
     bFunctions->PrintString("UI ONE");
+
               
     bFunctions->SetCursor_UI_Center();
+
     bFunctions->PrintString("UI TWO");
+
               
     bFunctions->SetCursor_UI_Right();
+
     bFunctions->PrintString("UI THREE");
 
 }
